Harden paper admission path with fail-closed gates and durable watch state.
Wire final pre-trade checks, creation/approval admission records, SQL CAS watch transitions, and isolated test fixtures so stale data cannot reach OrderIntent while the desk stays PAPER_TESTING_ONLY.

+++ backend/core/vendor_http.py
@@ -60,8 +60,8 @@ def retry_delay_seconds(exc: BaseException, *, attempt: int, base_delay: float) 
         header_wait = wait_seconds_from_headers(parsed)
         if header_wait is not None:
             return max(0.0, header_wait)
-        return fallback_429_wait(attempt)
-    return base_delay * (2**attempt)
+        return float(fallback_429_wait(attempt))
+    return float(base_delay * (2**attempt))
 
 
 async def get_with_retry(
